+++ packages/leva/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Patch Changes
 
+- 4f51de7: Switch from tinycolor to colord
 - dd17591: chore: update to stitches 0.2.0
 - d5a5dd2: - Fix: prevent selection on Firefox when closing the Color picker.
   - Fix: add suffix to Number plugin initial value.